corrected register logic in auth

+++ src/controllers/auth.controller.js
@@ -8,7 +8,29 @@ async function register(req,res){
     const {username,email,password}=req.body;
     const hashPassword=await bcrypt.hash(password,10);
     await pool.query("INSERT INTO users(email,username,password_hash) VALUES($1,$2,$3)",[email,username,hashPassword]);
-    res.status(201).json({message:'Registration Successful'});
+    const user=await pool.query('select user_id from users where email=$1',[email]);
+    const accessToken=jwt.sign({
+      id:user.rows[0].user_id
+    },process.env.JWT_SECRET,
+    {expiresIn:"15m"});
+
+    const refreshToken=jwt.sign({
+      id:user.rows[0].user_id
+    },process.env.JWT_SECRET,
+    {expiresIn:"1d"});
+    await pool.query('insert into refresh_tokens(user_id,token) values($1,$2)',[user.rows[0].user_id,refreshToken]);
+
+    
+    res.cookie('refreshToken',refreshToken,{
+    sameSite: process.env.NODE_ENV === 'production' ? 'none' : 'lax',
+    secure: process.env.NODE_ENV === 'production',
+    httpOnly:true,
+    maxAge:1*24*60*60000
+  });
+
+  res.status(201).json({message:'Registration Successful',
+      token:accessToken
+    });
   }catch(error){
     res.status(500).json({message:'Registration failed',error:error.message});
   }
@@ -38,7 +60,7 @@ async function login(req,res){
   const refreshToken=jwt.sign(
     {user:user.rows[0].user_id},
     process.env.JWT_SECRET,
-    {expiresIn:"7d"}
+    {expiresIn:"1d"}
   )
   await pool.query('INSERT INTO refresh_tokens (user_id,token) VALUES ($1,$2)',[user.rows[0].user_id,refreshToken]);
 
@@ -46,7 +68,7 @@ async function login(req,res){
     sameSite: process.env.NODE_ENV === 'production' ? 'none' : 'lax',
     secure: process.env.NODE_ENV === 'production',
     httpOnly:true,
-    maxAge:7*24*60*60000
+    maxAge:1*24*60*60000
   });
 
   res.json({message:'Successful login!',
@@ -63,8 +85,10 @@ try{
   const refreshToken=req.cookies.refreshToken;
   if(!refreshToken) return res.status(401).json({message:'cookie not found'});
   const payload=jwt.verify(refreshToken,process.env.JWT_SECRET);
+  //with jwt.verify() we check for 2 things: if the token has not expired and if it has not been tampered
 
   const tokenObj=await pool.query('SELECT token FROM refresh_tokens where token=$1',[refreshToken]);
+  
   
   if(payload && tokenObj.rowCount!=0){
     const accessToken=jwt.sign({id:payload.user},process.env.JWT_SECRET,{expiresIn:"15m"});
@@ -74,18 +98,21 @@ try{
     res.status(401).json({message:"invalid user"});
   }
 }catch(error){
-  res.json({message:'token expired'});
+  res.status(400).json({message:'token expired'});
 }  
   
   
 }
 async function logout(req,res){
  try{
+  if(!req.cookies.refreshToken){
+    return res.status(400).json({message:"token not found"});
+  }
   res.clearCookie("refreshToken");
   await pool.query('DELETE FROM refresh_tokens WHERE token=$1',[req.cookies.refreshToken]);
   res.json({message:'logout successful'});
  }catch(error){
-  res.status(401).json({message:error.message});
+  res.status(500).json({message:error.message});
  } 
   
 }
